requirements.txt
Django>=5.1.8,<5.2
djangorestframework>=3.16.0,<3.17
drf-spectacular>=0.28.0,<0.29
django-extensions>=4.1,<4.2
flake8>=7.1.1,<7.2
isort>=6.0.0,<6.1